Adding plugin decompilation
