+++ UniversalSearch.cs
@@ -1,29 +1,16 @@
 ﻿using McTools.Xrm.Connection;
-using Microsoft.Crm.Sdk.Messages;
 using Microsoft.VisualBasic.CompilerServices;
 using Microsoft.Xrm.Sdk;
-using Microsoft.Xrm.Sdk.Client;
-using Microsoft.Xrm.Sdk.Messages;
 using Microsoft.Xrm.Sdk.Metadata;
-using Microsoft.Xrm.Sdk.Organization;
-using Microsoft.Xrm.Sdk.Query;
-using MikeFactorial.XTB.Plugins.Xsd;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
-using System.Reflection;
-using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading;
 using System.Windows.Forms;
-using System.Xml;
-using System.Xml.Linq;
 using xrmtb.XrmToolBox.Controls;
 using XrmToolBox.Extensibility;
 using XrmToolBox.Extensibility.Args;
@@ -39,12 +26,14 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
 
         public string HelpUrl => "https://mikefactorial.com/dynamics-365-universal-search-for-xrmtoolbox/";
 
+        private bool solutionsSorted = false;
         public UniversalSearch()
         {
             InitializeComponent();
-            EntitiesListView.SolutionsDropDown.ComboBox.DropDownWidth = 400;
-            EntitiesListView.SolutionsDropDown.ComboBox.DropDownHeight = 300;
-            EntitiesListView.SplitContainerToolbar.SplitterDistance = (int)(EntitiesListView.SplitContainerToolbar.ClientSize.Width * .4);
+            this.SolutionDropDownComboBox.DropDownWidth = 400;
+            this.SolutionDropDownComboBox.DropDownHeight = 300;
+            this.SplitContainerToolbar.SplitterDistance = (int)(this.SplitContainerToolbar.ClientSize.Width * .4);
+            this.SolutionDropDownComboBox.DataSourceChanged += SolutionsDropDown_DataSourceChanged;
             EntitiesListView.SortList(0, SortOrder.Ascending);
         }
 
@@ -53,6 +42,54 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
         delegate void AddMetadataTabCallback(string entityName, List<MetadataSearchResult> results);
         delegate void AddSolutionTabCallback(string directoryName, List<SolutionSearchResult> results);
 
+        public CheckBox CheckBoxCheckAllNone
+        {
+            get
+            {
+                return this.Controls.Find("checkBoxCheckAllNone", true).ToList().FirstOrDefault() as CheckBox;
+            }
+        }
+        public ListView EntityListView
+        {
+            get
+            {
+                return this.Controls.Find("ListViewMain", true).ToList().FirstOrDefault() as ListView;
+            }
+        }
+        public SplitContainer SplitContainerToolbar
+        {
+            get
+            {
+                return this.Controls.Find("splitContainerToolbar", true).ToList().FirstOrDefault() as SplitContainer;
+            }
+        }
+        public SolutionsDropdownControl SolutionsDropDown
+        {
+            get
+            {
+                return this.Controls.Find("solutionsDropDown", true).ToList().FirstOrDefault() as SolutionsDropdownControl;
+            }
+        }
+        public ComboBox SolutionDropDownComboBox
+        {
+            get
+            {
+                return this.Controls.Find("comboSolutions", true).ToList().FirstOrDefault() as ComboBox;
+            }
+        }
+        private void SolutionsDropDown_DataSourceChanged(object sender, EventArgs e)
+        {
+            sortSolutionList();
+            if (this.SolutionsDropDown != null && this.SolutionDropDownComboBox.DataSource != null)
+            {
+                var defaultSolution = ((List<ListDisplayItem>)this.SolutionDropDownComboBox.DataSource).FirstOrDefault(i => i.Name.ToLower() == "default");
+                this.SolutionDropDownComboBox.SelectedItem = defaultSolution;
+            }
+        }
+        protected override void OnLoad(EventArgs e)
+        {
+            base.OnLoad(e);
+        }
         /// <summary>
         /// This event occurs when the connection has been updated in XrmToolBox
         /// </summary>
@@ -71,816 +108,6 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
             }
             EntitiesListView.Service = newService;
             base.UpdateConnection(newService, detail, actionName, parameter);
-        }
-
-        private void LoadMetadata(List<EntityMetadata> selectedEntities)
-        {
-            this.tabControl1.TabPages.Clear();
-            WorkAsync(new WorkAsyncInfo
-            {
-                IsCancelable = true,
-                Message = "Searching...",
-                AsyncArgument = selectedEntities,
-                Work = (worker, args) =>
-                {
-                    selectedEntities = selectedEntities.OrderBy(e => e.LogicalName).ToList();
-                    long recordCount = 0;
-                    int entityCount = 0;
-                    List<EntityMetadata> nonSelectedEntities = new List<EntityMetadata>();
-                    int errors = 0;
-                    for (int i = 0; i < selectedEntities.Count; i++)
-                    {
-                        try
-                        {
-                            string textToSearch = searchTextBox.Text;
-                            EntityMetadata selectedEntity = selectedEntities[i];
-                            double percentage = (double)(i + 1) / (double)selectedEntities.Count;
-                            percentage *= (double)100;
-
-                            List<AttributeMetadata> attsToSearch = null;
-                            worker.ReportProgress(Convert.ToInt32(percentage), $"Searching {selectedEntity.LogicalName} metadata.");
-                            if (searchAttributes.Checked)
-                            {
-                                // Retrieve the attribute metadata
-                                var req = new RetrieveEntityRequest()
-                                {
-                                    LogicalName = selectedEntity.LogicalName,
-                                    EntityFilters = EntityFilters.Attributes,
-                                    RetrieveAsIfPublished = true
-                                };
-                                var resp = (RetrieveEntityResponse)this.Service.Execute(req);
-                                if (worker.CancellationPending)
-                                {
-                                    return;
-                                }
-                                attsToSearch = resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.String || meta.AttributeType == AttributeTypeCode.Memo) && meta.IsValidForRead.Value).ToList();
-                            }
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            Guid guidValue;
-                            if (Guid.TryParse(this.searchTextBox.Text, out guidValue))
-                            {
-                                textToSearch = guidValue.ToString();
-                            }
-
-                            EntityMetadata relationshipMetadata = null;
-                            if (attsToSearch != null)
-                            {
-                                attsToSearch = attsToSearch.OrderBy(a => a.LogicalName).ToList();
-                            }
-                            if (searchRelationships.Checked)
-                            {
-                                RetrieveEntityRequest request = new RetrieveEntityRequest();
-                                request.EntityFilters = EntityFilters.Relationships;
-                                request.LogicalName = selectedEntity.LogicalName;
-
-                                var relationshipResponse = (RetrieveEntityResponse)this.Service.Execute(request);
-                                relationshipMetadata = relationshipResponse.EntityMetadata;
-                            }
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            var formsToSearch = new List<FormXml>();
-                            var fetchToSearch = new List<FetchXml>();
-                            var layoutToSearch = new List<LayoutXml>();
-                            if (searchFormsViews.Checked)
-                            {
-                                QueryExpression formQuery = new QueryExpression("systemform");
-                                formQuery.ColumnSet = new ColumnSet("formxml", "name");
-                                formQuery.Criteria = new FilterExpression(LogicalOperator.And);
-                                formQuery.Criteria.AddCondition("objecttypecode", ConditionOperator.Equal, selectedEntity.ObjectTypeCode.Value);
-                                var forms = this.Service.RetrieveMultiple(formQuery);
-                                foreach (var form in forms.Entities)
-                                {
-                                    if (form.Contains("formxml"))
-                                    {
-                                        var formObject = FormXml.Deserialize(form["formxml"].ToString());
-                                        if (form.Contains("name"))
-                                        {
-                                            formObject.FormName = form["name"].ToString();
-                                        }
-                                        else
-                                        {
-                                            formObject.FormName = "Unknown Form";
-                                        }
-                                        formsToSearch.Add(formObject);
-                                    }
-                                }
-
-                                if (worker.CancellationPending)
-                                {
-                                    return;
-                                }
-                                QueryExpression userQueryQuery = new QueryExpression("savedquery");
-                                userQueryQuery.ColumnSet = new ColumnSet("fetchxml", "layoutxml", "columnsetxml", "name");
-                                userQueryQuery.Criteria = new FilterExpression(LogicalOperator.And);
-                                userQueryQuery.Criteria.AddCondition("returnedtypecode", ConditionOperator.Equal, selectedEntity.ObjectTypeCode.Value);
-                                var userViews = this.Service.RetrieveMultiple(userQueryQuery);
-                                foreach (var view in userViews.Entities)
-                                {
-                                    if (view.Contains("fetchxml"))
-                                    {
-                                        var fetchObject = FetchXml.Deserialize(view["fetchxml"].ToString());
-                                        if (view.Contains("name"))
-                                        {
-                                            fetchObject.ViewName = view["name"].ToString();
-                                        }
-                                        else
-                                        {
-                                            fetchObject.ViewName = "Unknown View";
-                                        }
-                                        fetchToSearch.Add(fetchObject);
-                                    }
-
-                                    if (view.Contains("layoutxml"))
-                                    {
-                                        var layoutObject = LayoutXml.Deserialize(view["layoutxml"].ToString());
-                                        if (view.Contains("name"))
-                                        {
-                                            layoutObject.ViewName = view["name"].ToString();
-                                        }
-                                        else
-                                        {
-                                            layoutObject.ViewName = "Unknown View";
-                                        }
-                                        layoutToSearch.Add(layoutObject);
-                                    }
-                                }
-                            }
-
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            List<MetadataSearchResult> results = new List<MetadataSearchResult>();
-
-                            string itemIdentifier = "";
-                            this.searchMetadataObject(selectedEntity, string.Empty, "Entity", selectedEntity, itemIdentifier, textToSearch, ref results);
-                            if (formsToSearch.Count > 0)
-                            {
-                                this.searchMetadataObject(selectedEntity, "forms", "Form", formsToSearch, itemIdentifier, textToSearch, ref results);
-                            }
-                            if (fetchToSearch.Count > 0)
-                            {
-                                this.searchMetadataObject(selectedEntity, "views", "View Query", fetchToSearch, itemIdentifier, textToSearch, ref results);
-                            }
-                            if (layoutToSearch.Count > 0)
-                            {
-                                this.searchMetadataObject(selectedEntity, "views", "View Layout", layoutToSearch, itemIdentifier, textToSearch, ref results);
-                            }
-                            if (attsToSearch != null)
-                            {
-                                this.searchMetadataObject(selectedEntity, "fields", "Attribute", attsToSearch, itemIdentifier, textToSearch, ref results);
-                            }
-                            if (relationshipMetadata != null)
-                            {
-                                this.searchMetadataObject(selectedEntity, "relationships", "Relationship", relationshipMetadata, itemIdentifier, textToSearch, ref results);
-                            }
-
-                            recordCount += results.Count;
-
-                            AddMetadataResultTab(selectedEntity.LogicalName, results);
-
-                            entityCount++;
-                        }
-                        catch (Exception e)
-                        {
-                            errors++;
-                            try
-                            {
-                                this.LogError(e.ToString());
-                            }
-                            catch { }
-                            worker.ReportProgress(0, $"Search Completed with Errors. Error Message: " + e.Message);
-                        }
-                    }
-                    if (recordCount == 0)
-                    {
-                        worker.ReportProgress(0, $"Search Complete. No metadata was found. Make sure you've selected the correct entities to search and that you are using * for wildcard searches otherwise the exact text will be matched.");
-                    }
-                    else
-                    {
-                        worker.ReportProgress(0, $"Search Complete. Found {recordCount} instances in {entityCount} entities with {errors} error(s).");
-                    }
-
-                },
-                ProgressChanged = (args) =>
-                {
-                    SetWorkingMessage(args.UserState?.ToString());
-                    SendMessageToStatusBar(this, new StatusBarMessageEventArgs(args.ProgressPercentage, args.UserState.ToString().Replace("\r\n", " ")));
-                },
-                PostWorkCallBack = (args) =>
-                {
-                    this.searchLocationList.Enabled = true;
-                    this.btnFind.Text = "Search";
-                    if (args.Error != null)
-                    {
-                        MessageBox.Show(args.Error.Message, "Uh oh.", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                        return;
-                    }
-                }
-            });
-        }
-
-        private void LoadSolution(object selectedSolution)
-        {
-            this.tabControl1.TabPages.Clear();
-            WorkAsync(new WorkAsyncInfo
-            {
-                IsCancelable = true,
-                Message = $"Exporting {((ListDisplayItem)selectedSolution).ToString()}...",
-                AsyncArgument = selectedSolution,
-                Work = (worker, args) =>
-                {
-                    long recordCount = 0;
-                    int fileCount = 0;
-                    int errors = 0;
-                    string solutionName = ((ListDisplayItem)selectedSolution).Name;
-                    string solutionZipFileName = solutionName + ".zip";
-                    if (alwaysGetLatestSolutionCheckBox.Checked || !File.Exists(solutionZipFileName) || !Directory.Exists(solutionName))
-                    {
-                        // Export a solution
-                        ExportSolutionRequest exportSolutionRequest = new ExportSolutionRequest();
-                        exportSolutionRequest.Managed = false;
-                        exportSolutionRequest.SolutionName = solutionName;
-
-                        ExportSolutionResponse exportSolutionResponse =
-                           (ExportSolutionResponse)this.Service.Execute(exportSolutionRequest);
-
-                        byte[] exportXml = exportSolutionResponse.ExportSolutionFile;
-                        if (worker.CancellationPending)
-                        {
-                            return;
-                        }
-
-                        if (File.Exists(solutionZipFileName))
-                        {
-                            File.Delete(solutionZipFileName);
-                        }
-                        File.WriteAllBytes(solutionZipFileName, exportXml);
-                        if (Directory.Exists(solutionName))
-                        {
-                            Directory.Delete(solutionName, true);
-                        }
-                        if (worker.CancellationPending)
-                        {
-                            return;
-                        }
-                        ZipFile.ExtractToDirectory(solutionZipFileName, solutionName);
-                        if (worker.CancellationPending)
-                        {
-                            return;
-                        }
-                    }
-                    var files = Directory.GetFiles(solutionName, "*.*", SearchOption.AllDirectories);
-                    for (int i = 0; i < files.Length; i++)
-                    {
-                        try
-                        {
-                            string textToSearch = searchTextBox.Text;
-                            double percentage = (double)(i + 1) / (double)files.Length;
-                            percentage *= (double)100;
-                            Thread.Sleep(500);
-                            worker.ReportProgress(Convert.ToInt32(percentage), $"Searching {files[i]}.");
-
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            List<SolutionSearchResult> results = new List<SolutionSearchResult>();
-
-                            if (files[i].EndsWith(".msapp"))
-                            {
-                                var msappUnpackPath = files[i].Replace(".msapp", "");
-                                if (Directory.Exists(msappUnpackPath))
-                                {
-                                    Directory.Delete(msappUnpackPath, true);
-                                }
-
-                                var canvasDir = Directory.CreateDirectory(msappUnpackPath);
-                                ZipFile.ExtractToDirectory(files[i], msappUnpackPath);
-
-                                var appFiles = Directory.GetFiles(msappUnpackPath, "*.*", SearchOption.AllDirectories);
-                                foreach(var appFile in appFiles)
-                                {
-                                    searchSolutionObject(appFile, textToSearch, ref results);
-                                }
-                            }
-                            else
-                            {
-                                searchSolutionObject(files[i], textToSearch, ref results);
-                            }
-                            recordCount += results.Count;
-                            AddSolutionResultTab(Path.GetFileName(files[i]), results);
-                            fileCount++;
-                        }
-                        catch (Exception e)
-                        {
-                            errors++;
-                            try
-                            {
-                                this.LogError(e.ToString());
-                            }
-                            catch { }
-                            worker.ReportProgress(0, $"Search Completed with Errors. Error Message: " + e.Message);
-                        }
-                    }
-                    if (recordCount == 0)
-                    {
-                        worker.ReportProgress(0, $"Search Complete. No files were found. Make sure you've selected the correct solution to search and that you are using * for wildcard searches otherwise the exact text will be matched.");
-                    }
-                    else
-                    {
-                        worker.ReportProgress(0, $"Search Complete. Found {recordCount} instances in {fileCount} files with {errors} error(s).");
-                    }
-
-                },
-                ProgressChanged = (args) =>
-                {
-                    SetWorkingMessage(args.UserState?.ToString());
-                    SendMessageToStatusBar(this, new StatusBarMessageEventArgs(args.ProgressPercentage, args.UserState.ToString().Replace("\r\n", " ")));
-                },
-                PostWorkCallBack = (args) =>
-                {
-                    this.searchLocationList.Enabled = true;
-                    this.btnFind.Text = "Search";
-                    if (args.Error != null)
-                    {
-                        MessageBox.Show(args.Error.Message, "Uh oh.", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                        return;
-                    }
-                }
-            });
-        }
-
-        private void searchSolutionObject(string filePath, string searchText, ref List<SolutionSearchResult> results)
-        {
-            //Format xml documents first
-            if (filePath.Contains(".xml"))
-            {
-                // Load the XML file
-                XDocument doc = XDocument.Load(filePath);
-                // Format the XML with indents and line breaks
-                string formattedXml = doc.ToString();
-
-                // Write the formatted XML to a new file
-                File.WriteAllText(filePath, formattedXml);
-            }
-            else if (filePath.Contains(".json"))
-            {
-                // Read the JSON file into a string
-                string json = File.ReadAllText(filePath);
-                // Parse the string into a JObject
-                JObject parsedJson = JObject.Parse(json);
-                // Format the JSON with indents and line breaks
-                string formattedJson = parsedJson.ToString(Newtonsoft.Json.Formatting.Indented);
-                // Write the formatted JSON to a new file
-                File.WriteAllText(filePath, formattedJson);
-            }
-            var regEx = "^" + Regex.Escape(searchText).Replace("\\*", ".*") + "$";
-            var lines = File.ReadLines(filePath).ToList();
-            for (int i = 0; i < lines.Count; i++)
-            {
-                if ((!solutionMatchCaseCheckBox.Checked && Regex.IsMatch(lines[i], regEx, RegexOptions.IgnoreCase)) || (solutionMatchCaseCheckBox.Checked && Regex.IsMatch(lines[i], regEx)))
-                {
-                    results.Add(new SolutionSearchResult() { FileLink = filePath, LineNumber = (i+1), FilePath = filePath, Value = lines[i] });
-                }
-            }
-        }
-        private void searchMetadataObject(EntityMetadata entity, string linkType, string metadataType, object searchObject, string itemIdentifier, string searchText, ref List<MetadataSearchResult> results)
-        {
-            if (searchObject == null) return;
-            itemIdentifier = this.BuildItemIdentifier(searchObject, itemIdentifier);
-            var listObject = searchObject as System.Collections.IList;
-            if (listObject != null)
-            {
-                foreach (var listItem in listObject)
-                {
-                    searchMetadataObject(entity, linkType, metadataType, listItem, itemIdentifier, searchText, ref results);
-                }
-            }
-            else
-            {
-                Type objType = searchObject.GetType();
-                PropertyInfo[] properties = objType.GetProperties(BindingFlags.Instance | BindingFlags.Public).Where(p => p.CanRead).ToArray();
-                foreach (PropertyInfo property in properties)
-                {
-                    object propValue = property.GetValue(searchObject, null);
-                    var elems = propValue as System.Collections.IList;
-                    if (elems != null)
-                    {
-                        foreach (var item in elems)
-                        {
-                            searchMetadataObject(entity, linkType, metadataType, item, itemIdentifier, searchText, ref results);
-                        }
-                    }
-                    else
-                    {
-                        // This will not cut-off System.Collections because of the first check
-                        if (property.PropertyType.Assembly == objType.Assembly)
-                        {
-                            searchMetadataObject(entity, linkType, metadataType, propValue, itemIdentifier, searchText, ref results);
-                        }
-                        else
-                        {
-                            if (!matchCaseMetadata.Checked)
-                            {
-                                if (Regex.IsMatch(property.Name, WildCardToRegular(searchText), RegexOptions.IgnoreCase) || (!string.IsNullOrEmpty(propValue?.ToString()) && Regex.IsMatch(propValue?.ToString(), WildCardToRegular(searchText), RegexOptions.IgnoreCase)))
-                                {
-                                    results.Add(new MetadataSearchResult()
-                                    {
-                                        Link = this.BuildLinkForMetadataItem(entity, linkType),
-                                        Property = property.Name,
-                                        Value = propValue?.ToString(),
-                                        Item = itemIdentifier,
-                                        Type = objType.ToString().Split('.').Last(),
-                                        Metadata = metadataType
-                                    });
-                                }
-                            }
-                            else if (Regex.IsMatch(property.Name, WildCardToRegular(searchText)) || (!string.IsNullOrEmpty(propValue?.ToString()) && Regex.IsMatch(propValue?.ToString(), WildCardToRegular(searchText))))
-                            {
-                                results.Add(new MetadataSearchResult()
-                                {
-                                    Link = this.BuildLinkForMetadataItem(entity, linkType),
-                                    Property = property.Name,
-                                    Value = propValue?.ToString(),
-                                    Item = itemIdentifier,
-                                    Type = objType.ToString().Split('.').Last(),
-                                    Metadata = metadataType
-                                });
-                            }
-                        }
-                    }
-                }
-            }
-            return;
-        }
-
-        private string BuildItemIdentifier(object searchObject, string currentIdentifier)
-        {
-            if (searchObject is EntityMetadata)
-            {
-                return ((EntityMetadata)searchObject).DisplayName?.UserLocalizedLabel?.Label;
-            }
-            else if (searchObject is AttributeMetadata)
-            {
-                return ((AttributeMetadata)searchObject).DisplayName?.UserLocalizedLabel?.Label;
-            }
-            else if (searchObject is OneToManyRelationshipMetadata)
-            {
-                return $"{((OneToManyRelationshipMetadata)searchObject)?.ReferencedEntity} ({((OneToManyRelationshipMetadata)searchObject)?.ReferencedAttribute}) - {((OneToManyRelationshipMetadata)searchObject)?.ReferencingEntity} ({((OneToManyRelationshipMetadata)searchObject)?.ReferencingAttribute})";
-            }
-            else if (searchObject is ManyToManyRelationshipMetadata)
-            {
-                return $"{((ManyToManyRelationshipMetadata)searchObject)?.Entity1LogicalName} ({((ManyToManyRelationshipMetadata)searchObject)?.Entity1IntersectAttribute}) - {((ManyToManyRelationshipMetadata)searchObject)?.Entity2LogicalName} ({((ManyToManyRelationshipMetadata)searchObject)?.Entity2IntersectAttribute})";
-            }
-            else if (searchObject is FormXml)
-            {
-                return ((FormXml)searchObject).FormName;
-            }
-            else if (searchObject is FetchXml)
-            {
-                return ((FetchXml)searchObject).ViewName;
-            }
-            else if (searchObject is LayoutXml)
-            {
-                return ((LayoutXml)searchObject).ViewName;
-            }
-            return currentIdentifier;
-        }
-        private string BuildLinkForMetadataItem(EntityMetadata entity, string type)
-        {
-            return $"https://make.powerapps.com/environments/{ConnectionDetail.Organization}/entities/{entity.MetadataId.ToString()}/{entity.LogicalName}#{type}";
-        }
-
-        private void LoadData(List<EntityMetadata> selectedEntities)
-        {
-            this.tabControl1.TabPages.Clear();
-            WorkAsync(new WorkAsyncInfo
-            {
-                IsCancelable = true,
-                Message = "Searching...",
-                AsyncArgument = selectedEntities,
-                Work = (worker, args) =>
-                {
-                    selectedEntities = selectedEntities.OrderBy(e => e.LogicalName).ToList();
-                    long recordCount = 0;
-                    int entityCount = 0;
-                    List<EntityMetadata> nonSelectedEntities = new List<EntityMetadata>();
-                    Dictionary<Guid, string> matchedLookups = new Dictionary<Guid, string>();
-                    int errors = 0;
-                    for (int i = 0; i < selectedEntities.Count; i++)
-                    {
-                        try
-                        {
-                            EntityMetadata selectedEntity = selectedEntities[i];
-                            double percentage = (double)(i + 1) / (double)selectedEntities.Count;
-                            percentage = percentage * (double)100;
-
-                            worker.ReportProgress(Convert.ToInt32(percentage), $"Searching for {selectedEntity.LogicalName} records.");
-                 
-                            // Retrieve the attribute metadata
-                            var req = new RetrieveEntityRequest()
-                            {
-                                LogicalName = selectedEntity.LogicalName,
-                                EntityFilters = EntityFilters.Attributes,
-                                RetrieveAsIfPublished = true
-                            };
-                            var resp = (RetrieveEntityResponse)this.Service.Execute(req);
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            List<AttributeMetadata> attsToSearch = resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.String || meta.AttributeType == AttributeTypeCode.Memo) && meta.IsValidForRead.Value).ToList();
-
-
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            Guid guidValue;
-                            if (Guid.TryParse(this.searchTextBox.Text, out guidValue))
-                            {
-                                attsToSearch.AddRange(resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.Lookup || meta.AttributeType == AttributeTypeCode.Owner || meta.AttributeType == AttributeTypeCode.Uniqueidentifier || meta.AttributeType == AttributeTypeCode.Customer) && meta.IsValidForRead.Value).ToList());
-                            }
-                            else if (searchLookupText.Checked)
-                            {
-                                attsToSearch.AddRange(resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.Lookup || meta.AttributeType == AttributeTypeCode.Owner || meta.AttributeType == AttributeTypeCode.Customer) && meta.IsValidForRead.Value).ToList());
-                            }
-
-                            long intValue;
-                            if (long.TryParse(this.searchTextBox.Text, out intValue))
-                            {
-                                attsToSearch.AddRange(resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.BigInt || meta.AttributeType == AttributeTypeCode.Integer || meta.AttributeType.Value == AttributeTypeCode.Picklist || meta.AttributeType.Value == AttributeTypeCode.State || meta.AttributeType.Value == AttributeTypeCode.Status) && meta.IsValidForRead.Value).ToList());
-                            }
-                            else
-                            {
-                                if (searchOptionSetText.Checked)
-                                {
-                                    attsToSearch.AddRange(resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.Picklist) && meta.IsValidForRead.Value));
-                                }
-                                intValue = long.MinValue;
-                            }
-
-                            double doubleValue;
-                            if (double.TryParse(this.searchTextBox.Text, out doubleValue))
-                            {
-                                attsToSearch.AddRange(resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.Money || meta.AttributeType == AttributeTypeCode.Decimal || meta.AttributeType.Value == AttributeTypeCode.Double) && meta.IsValidForRead.Value).ToList());
-                            }
-                            DateTime dateTimeValue;
-                            if (DateTime.TryParse(this.searchTextBox.Text, CultureInfo.InvariantCulture, DateTimeStyles.AssumeLocal, out dateTimeValue))
-                            {
-                                attsToSearch.AddRange(resp.EntityMetadata.Attributes.Where(meta => (meta.AttributeType == AttributeTypeCode.DateTime) && meta.IsValidForRead.Value).ToList());
-                            }
-
-                            QueryExpression query = new QueryExpression(selectedEntity.LogicalName);
-                            query.ColumnSet = new ColumnSet();
-                            FilterExpression fe = new FilterExpression(LogicalOperator.Or);
-                            query.ColumnSet.AddColumn(selectedEntity.PrimaryIdAttribute);
-                            AttributeMetadata primaryNameAttribute = resp.EntityMetadata.Attributes.FirstOrDefault(a => a.LogicalName == selectedEntity.PrimaryNameAttribute);
-                            if (primaryNameAttribute != null)
-                            {
-                                if (primaryNameAttribute.IsValidForRead.Value && string.IsNullOrEmpty(primaryNameAttribute.AttributeOf))
-                                {
-                                    query.ColumnSet.AddColumn(primaryNameAttribute.LogicalName);
-                                }
-                                fe.AddCondition(primaryNameAttribute.LogicalName, ConditionOperator.Like, this.searchTextBox.Text.Replace("*", "%"));
-                            }
-                            if (guidValue != Guid.Empty)
-                            {
-                                fe.AddCondition(selectedEntity.PrimaryIdAttribute, ConditionOperator.Equal, guidValue);
-                            }
-                            attsToSearch = attsToSearch.OrderBy(a => a.LogicalName).ToList();
-                            bool conditionAdded = false;
-                            EntityMetadata relationshipMetadata = null;
-                            if (searchLookupText.Checked)
-                            {
-                                RetrieveEntityRequest request = new RetrieveEntityRequest();
-                                request.EntityFilters = EntityFilters.Relationships;
-                                request.LogicalName = selectedEntity.LogicalName;
-
-                                var relationshipResponse = (RetrieveEntityResponse)this.Service.Execute(request);
-                                relationshipMetadata = relationshipResponse.EntityMetadata;
-                            }
-
-                            foreach (AttributeMetadata att in attsToSearch)
-                            {
-                                if (att.LogicalName != selectedEntity.PrimaryNameAttribute && att.LogicalName != selectedEntity.PrimaryIdAttribute)
-                                {
-                                    if (string.IsNullOrEmpty(att.AttributeOf))
-                                    {
-                                        query.ColumnSet.AddColumn(att.LogicalName);
-                                        if (att.AttributeType == AttributeTypeCode.Lookup || att.AttributeType == AttributeTypeCode.Owner || att.AttributeType == AttributeTypeCode.Uniqueidentifier || att.AttributeType == AttributeTypeCode.Customer)
-                                        {
-                                            if (searchLookupText.Checked && (att.AttributeType == AttributeTypeCode.Lookup || att.AttributeType == AttributeTypeCode.Owner || att.AttributeType == AttributeTypeCode.Customer) && guidValue == Guid.Empty)
-                                            {
-                                                //Here we are going to search for the picklist display value rather than the value
-                                                LookupAttributeMetadata lookup = (LookupAttributeMetadata)att;
-                                                OneToManyRelationshipMetadata relationship = relationshipMetadata.ManyToOneRelationships.FirstOrDefault(r => r.ReferencingAttribute == lookup.LogicalName);
-
-                                                if (relationship != null)
-                                                {
-                                                    QueryExpression lookupQuery = new QueryExpression(relationship.ReferencedEntity);
-                                                    EntityMetadata relatedEntityMetadata = selectedEntities.FirstOrDefault(m => m.LogicalName == relationship.ReferencedEntity);
-                                                    if (relatedEntityMetadata == null)
-                                                    {
-                                                        relatedEntityMetadata = nonSelectedEntities.FirstOrDefault(m => m.LogicalName == relationship.ReferencedEntity);
-                                                        if (relatedEntityMetadata == null)
-                                                        {
-                                                            RetrieveEntityRequest request = new RetrieveEntityRequest();
-                                                            request.EntityFilters = EntityFilters.Entity;
-                                                            request.LogicalName = relationship.ReferencedEntity;
-
-                                                            var relationshipResponse = (RetrieveEntityResponse)this.Service.Execute(request);
-                                                            nonSelectedEntities.Add(relationshipResponse.EntityMetadata);
-                                                            relatedEntityMetadata = nonSelectedEntities.FirstOrDefault(m => m.LogicalName == relationship.ReferencedEntity);
-                                                        }
-                                                    }
-                                                    FilterExpression filter = new FilterExpression();
-                                                    filter.AddCondition(relatedEntityMetadata.PrimaryNameAttribute, ConditionOperator.Like, this.searchTextBox.Text.Replace("*", "%"));
-                                                    lookupQuery.Criteria.AddFilter(filter);
-                                                    lookupQuery.ColumnSet = new ColumnSet(new string[] { relatedEntityMetadata.PrimaryIdAttribute, relatedEntityMetadata.PrimaryNameAttribute });
-
-                                                    EntityCollection parentEntities = this.Service.RetrieveMultiple(lookupQuery);
-                                                    foreach (var parentEntity in parentEntities.Entities)
-                                                    {
-                                                        fe.AddCondition(lookup.LogicalName, ConditionOperator.Equal, parentEntity[relatedEntityMetadata.PrimaryIdAttribute]);
-                                                        if(!matchedLookups.ContainsKey((Guid)parentEntity[relatedEntityMetadata.PrimaryIdAttribute]))
-                                                        {
-                                                            matchedLookups.Add((Guid)parentEntity[relatedEntityMetadata.PrimaryIdAttribute], (string)parentEntity[relatedEntityMetadata.PrimaryNameAttribute]);
-                                                        }
-                                                        conditionAdded = true;
-                                                    }
-                                                }
-                                            }
-                                            else
-                                            {
-                                                fe.AddCondition(att.LogicalName, ConditionOperator.Equal, guidValue);
-                                                conditionAdded = true;
-                                            }
-                                        }
-                                        else if (att.AttributeType == AttributeTypeCode.BigInt)
-                                        {
-                                            fe.AddCondition(att.LogicalName, ConditionOperator.Equal, intValue);
-                                            conditionAdded = true;
-                                        }
-                                        else if (att.AttributeType == AttributeTypeCode.Integer || att.AttributeType.Value == AttributeTypeCode.Picklist || att.AttributeType.Value == AttributeTypeCode.State || att.AttributeType.Value == AttributeTypeCode.Status)
-                                        {
-                                            if (searchOptionSetText.Checked && (att.AttributeType.Value == AttributeTypeCode.Picklist && intValue == long.MinValue))
-                                            {
-                                                //Here we are going to search for the picklist display value rather than the value
-                                                PicklistAttributeMetadata pick = (PicklistAttributeMetadata)att;
-                                                if (pick.OptionSet != null)
-                                                {
-                                                    foreach (var option in pick.OptionSet.Options)
-                                                    {
-                                                        if (option.Label != null && option.Label.UserLocalizedLabel != null && option.Value != null)
-                                                        {
-                                                            if (!matchCaseCheckBox.Checked)
-                                                            {
-                                                                if (Regex.IsMatch(option.Label.UserLocalizedLabel.Label, WildCardToRegular(this.searchTextBox.Text), RegexOptions.IgnoreCase))
-                                                                {
-                                                                    fe.AddCondition(att.LogicalName, ConditionOperator.Equal, (Int32)option.Value.Value);
-                                                                    conditionAdded = true;
-                                                                }
-                                                            }
-                                                            else if (Regex.IsMatch(option.Label.UserLocalizedLabel.Label, WildCardToRegular(this.searchTextBox.Text)))
-                                                            {
-                                                                fe.AddCondition(att.LogicalName, ConditionOperator.Equal, (Int32)option.Value.Value);
-                                                                conditionAdded = true;
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                            else
-                                            {
-                                                fe.AddCondition(att.LogicalName, ConditionOperator.Equal, (Int32)intValue);
-                                                conditionAdded = true;
-                                            }
-                                        }
-                                        else if (att.AttributeType.Value == AttributeTypeCode.DateTime)
-                                        {
-                                            fe.AddCondition(att.LogicalName, ConditionOperator.On, dateTimeValue.Date);
-                                            conditionAdded = true;
-                                        }
-                                        else if (att.AttributeType.Value == AttributeTypeCode.Double)
-                                        {
-                                            fe.AddCondition(att.LogicalName, ConditionOperator.Equal, doubleValue);
-                                            conditionAdded = true;
-                                        }
-                                        else if (att.AttributeType == AttributeTypeCode.Money || att.AttributeType == AttributeTypeCode.Decimal)
-                                        {
-                                            fe.AddCondition(att.LogicalName, ConditionOperator.Equal, (decimal)doubleValue);
-                                            conditionAdded = true;
-                                        }
-                                        else
-                                        {
-                                            fe.AddCondition(att.LogicalName, ConditionOperator.Like, this.searchTextBox.Text.Replace("*", "%"));
-                                            conditionAdded = true;
-                                        }
-                                    }
-                                }
-                            }
-                            query.Criteria.AddFilter(fe);
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            if (worker.CancellationPending)
-                            {
-                                return;
-                            }
-                            if (conditionAdded)
-                            {
-                                EntityCollection results = Service.RetrieveMultiple(query);
-                                if (results.Entities != null && results.Entities.Count > 0)
-                                {
-                                    //Here we are going to replace the integer value with the display value for the picklist we returned in the results and convert UTC date/time to local for the results to display correctly..
-                                    UpdateDisplayTexts(attsToSearch, intValue, matchedLookups, guidValue, results);
-
-                                    if (this.matchCaseCheckBox.Checked || dateTimeValue != DateTime.MinValue)
-                                    {
-                                        for (int j = 0; j < results.Entities.Count; j++)
-                                        {
-                                            Entity e = results.Entities[j];
-                                            if (this.matchCaseCheckBox.Checked && !e.Attributes.Any(a => (a.Value != null && LikeOperator.LikeString(a.Value.ToString(), searchTextBox.Text, Microsoft.VisualBasic.CompareMethod.Binary))))
-                                            {
-                                                results.Entities.RemoveAt(j);
-                                                j--;
-                                            }
-                                            else if (dateTimeValue != DateTime.MinValue)
-                                            {
-                                                bool found = false;
-                                                foreach (var att in attsToSearch.Where(a => a.AttributeType != null && a.AttributeType.Value == AttributeTypeCode.DateTime))
-                                                {
-                                                    if(results.Entities[j].Contains(att.LogicalName) && ((DateTime)results.Entities[j][att.LogicalName]).Date == dateTimeValue.Date)
-                                                    {
-                                                        found = true;
-                                                        break;
-                                                    }
-                                                }
-                                                if (!found)
-                                                {
-                                                    results.Entities.RemoveAt(j);
-                                                    j--;
-                                                }
-                                            }
-                                        }
-                                        recordCount += results.Entities.Count;
-                                        entityCount++;
-                                        AddRecordResultTab(results);
-                                    }
-                                    else
-                                    {
-                                        recordCount += results.Entities.Count;
-                                        entityCount++;
-                                        AddRecordResultTab(results);
-                                    }
-                                }
-                            }
-                        }
-                        catch (Exception e)
-                        {
-                            errors++;
-                            try
-                            {
-                                this.LogError(e.ToString());
-                            }
-                            catch { }
-                            worker.ReportProgress(0, $"Search Completed with Errors. Error Message: " + e.Message);
-                        }
-                    }
-                    if (recordCount == 0)
-                    {
-                        worker.ReportProgress(0, $"Search Complete. No records were found. Make sure you've selected the correct entities to search and that you are using * for wildcard searches otherwise the exact text will be matched.");
-                    }
-                    else
-                    {
-                        worker.ReportProgress(0, $"Search Complete. Found {recordCount} records in {entityCount} entities with {errors} error(s).");
-                    }
-
-                },
-                ProgressChanged = (args) =>
-                {
-                    SetWorkingMessage(args.UserState?.ToString());
-                    SendMessageToStatusBar(this, new StatusBarMessageEventArgs(args.ProgressPercentage, args.UserState.ToString().Replace("\r\n", " ")));
-                },
-                PostWorkCallBack = (args) =>
-                {
-                    this.searchLocationList.Enabled = true;
-                    this.btnFind.Text = "Search";
-                    if (args.Error != null)
-                    {
-                        MessageBox.Show(args.Error.Message, "Uh oh.", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                        return;
-                    }
-                }
-            });
         }
 
         private void UpdateDisplayTexts(List<AttributeMetadata> attsToSearch, long intValue, Dictionary<Guid, string> matchedLookups, Guid guidValue, EntityCollection results)
@@ -994,78 +221,6 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
             }
         }
 
-        private void AddSolutionResultTab(string directoryName, List<SolutionSearchResult> results)
-        {
-            if (results.Count > 0)
-            {
-                if (this.tabControl1.InvokeRequired)
-                {
-                    AddSolutionTabCallback d = new AddSolutionTabCallback(AddSolutionResultTab);
-                    this.Invoke(d, new object[] { directoryName, results });
-                }
-                else
-                {
-                    DataGridView gridData = new DataGridView();
-                    gridData.AllowUserToAddRows = false;
-                    gridData.AllowUserToDeleteRows = false;
-                    gridData.AllowUserToOrderColumns = true;
-                    gridData.AllowUserToResizeRows = false;
-                    gridData.AutoSizeColumnsMode = System.Windows.Forms.DataGridViewAutoSizeColumnsMode.AllCells;
-                    gridData.ColumnHeadersHeightSizeMode = System.Windows.Forms.DataGridViewColumnHeadersHeightSizeMode.AutoSize;
-                    gridData.Dock = System.Windows.Forms.DockStyle.Fill;
-                    gridData.Location = new System.Drawing.Point(8, 38);
-                    gridData.Margin = new System.Windows.Forms.Padding(8, 7, 8, 7);
-                    gridData.Name = "gridData";
-                    gridData.ReadOnly = true;
-                    gridData.RowHeadersVisible = false;
-                    gridData.Size = new System.Drawing.Size(1587, 1090);
-                    gridData.TabIndex = 0;
-                    gridData.CellDoubleClick += gridData_SolutionCellDoubleClick;
-                    gridData.DataSource = results;
-                    gridData.DataBindingComplete += GridData_DataBindingComplete;
-                    TabPage tabPage = new TabPage(directoryName);
-                    tabPage.Controls.Add(gridData);
-                    this.tabControl1.TabPages.Add(tabPage);
-                }
-            }
-        }
-
-        private void AddMetadataResultTab(string entityName, List<MetadataSearchResult> results)
-        {
-            if (results.Count > 0)
-            {
-                if (this.tabControl1.InvokeRequired)
-                {
-                    AddMetadataTabCallback d = new AddMetadataTabCallback(AddMetadataResultTab);
-                    this.Invoke(d, new object[] { entityName, results });
-                }
-                else
-                {
-                    DataGridView gridData = new DataGridView();
-                    gridData.AllowUserToAddRows = false;
-                    gridData.AllowUserToDeleteRows = false;
-                    gridData.AllowUserToOrderColumns = true;
-                    gridData.AllowUserToResizeRows = false;
-                    gridData.AutoSizeColumnsMode = System.Windows.Forms.DataGridViewAutoSizeColumnsMode.AllCells;
-                    gridData.ColumnHeadersHeightSizeMode = System.Windows.Forms.DataGridViewColumnHeadersHeightSizeMode.AutoSize;
-                    gridData.Dock = System.Windows.Forms.DockStyle.Fill;
-                    gridData.Location = new System.Drawing.Point(8, 38);
-                    gridData.Margin = new System.Windows.Forms.Padding(8, 7, 8, 7);
-                    gridData.Name = "gridData";
-                    gridData.ReadOnly = true;
-                    gridData.RowHeadersVisible = false;
-                    gridData.Size = new System.Drawing.Size(1587, 1090);
-                    gridData.TabIndex = 0;
-                    gridData.CellDoubleClick += gridData_CellDoubleClick;
-                    gridData.DataSource = results;
-                    gridData.DataBindingComplete += GridData_DataBindingComplete;
-                    TabPage tabPage = new TabPage(entityName);
-                    tabPage.Controls.Add(gridData);
-                    this.tabControl1.TabPages.Add(tabPage);
-                }
-            }
-        }
-
         private void gridData_SolutionCellDoubleClick(object sender, DataGridViewCellEventArgs e)
         {
             if (e.RowIndex >= 0)
@@ -1076,7 +231,6 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
                 Process.Start(fullFilePath);
             }
         }
-
 
         private void gridData_CellDoubleClick(object sender, DataGridViewCellEventArgs e)
         {
@@ -1147,38 +301,6 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
                 }
             }
         }
-        private void OpenMetadataReference(string url)
-        {
-            if (!string.IsNullOrEmpty(url))
-            {
-                Process.Start(url);
-            }
-        }
-        private void OpenEntityReference(EntityReference entref)
-        {
-            if (!string.IsNullOrEmpty(entref.LogicalName) && !entref.Id.Equals(Guid.Empty))
-            {
-                var url = ConnectionDetail.WebApplicationUrl;
-                if (string.IsNullOrEmpty(url))
-                {
-                    url = string.Concat(ConnectionDetail.ServerName, "/", ConnectionDetail.Organization);
-                    if (!url.ToLower().StartsWith("http"))
-                    {
-                        url = string.Concat("http://", url);
-                    }
-                }
-                url = string.Concat(url,
-                    url.EndsWith("/") ? "" : "/",
-                    "main.aspx?etn=",
-                    entref.LogicalName,
-                    "&pagetype=entityrecord&id=",
-                    entref.Id.ToString());
-                if (!string.IsNullOrEmpty(url))
-                {
-                    Process.Start(url);
-                }
-            }
-        }
 
         private void gridData_RecordDoubleClick(object sender, xrmtb.XrmToolBox.Controls.CRMRecordEventArgs e)
         {
@@ -1219,88 +341,22 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
             }
         }
 
-        private void findInRecords()
-        {
-            if (btnFind.Text.StartsWith("Search"))
-            {
-                if (this.EntitiesListView.CheckedEntities.Count == 0)
-                {
-                    MessageBox.Show("Please select at least one entity to search before continuing.", "No Entities Selected", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                }
-                else if (string.IsNullOrEmpty(searchTextBox.Text))
-                {
-                    MessageBox.Show("Please enter your search criteria in the search box. Use asterisks * to perform a wildcard search.", "No Search Criteria Entered", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                }
-                else
-                {
-                    searchLocationList.Enabled = false;
-                    btnFind.Text = "Cancel";
-                    LoadData(this.EntitiesListView.CheckedEntities);
-                }
-            }
-            else
-            {
-                CancelWorker(); // PluginBaseControl method that calls the Background Workers CancelAsync method.
-                btnFind.Text = "Search";
-                searchLocationList.Enabled = true;
-            }
-        }
-        private void findInSolution()
-        {
-            if (btnFind.Text.StartsWith("Search"))
-            {
-                if (this.EntitiesListView.SolutionsDropDown.ComboBox.SelectedItem == null)
-                {
-                    MessageBox.Show("Please select a solution to search before continuing.", "No Solution Selected", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                }
-                else if (string.IsNullOrEmpty(searchTextBox.Text))
-                {
-                    MessageBox.Show("Please enter your search criteria in the search box. Use asterisks * to perform a wildcard search", "No Search Criteria Entered", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                }
-                else
-                {
-                    btnFind.Text = "Cancel";
-                    searchLocationList.Enabled = false;
-                    LoadSolution(this.EntitiesListView.SolutionsDropDown.ComboBox.SelectedItem);
-                }
-            }
-            else
-            {
-                CancelWorker(); // PluginBaseControl method that calls the Background Workers CancelAsync method.
-                btnFind.Text = "Search";
-                searchLocationList.Enabled = true;
-            }
-        }
-
-        private void findInMetadata()
-        {
-            if (btnFind.Text.StartsWith("Search"))
-            {
-                if (this.EntitiesListView.CheckedEntities.Count == 0)
-                {
-                    MessageBox.Show("Please select at least one entity to search before continuing.", "No Entities Selected", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                }
-                else if (string.IsNullOrEmpty(searchTextBox.Text))
-                {
-                    MessageBox.Show("Please enter your search criteria in the search box. Use asterisks * to perform a wildcard search", "No Search Criteria Entered", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                }
-                else
-                {
-                    btnFind.Text = "Cancel";
-                    searchLocationList.Enabled = false;
-                    LoadMetadata(this.EntitiesListView.CheckedEntities);
-                }
-            }
-            else
-            {
-                CancelWorker(); // PluginBaseControl method that calls the Background Workers CancelAsync method.
-                btnFind.Text = "Search";
-                searchLocationList.Enabled = true;
-            }
-        }
         private void SearchLocationList_SelectedIndexChanged(object sender, System.EventArgs e)
         {
             this.updateSearchVisibility();
+        }
+        private void sortSolutionList()
+        {
+            if (this.SolutionDropDownComboBox.DataSource != null)
+            {
+                ((List<ListDisplayItem>)this.SolutionDropDownComboBox.DataSource).Sort(delegate (ListDisplayItem x, ListDisplayItem y)
+                {
+                    if (x.DisplayName == null && y.DisplayName == null) return 0;
+                    else if (x.DisplayName == null) return -1;
+                    else if (y.DisplayName == null) return 1;
+                    else return x.DisplayName.CompareTo(y.DisplayName);
+                });
+            }
         }
         private void updateSearchVisibility()
         {
@@ -1310,18 +366,25 @@ namespace MikeFactorial.XTB.Plugins.UniversalSearch
             resultsGroup.Text = (searchLocationList.SelectedIndex == 1) ? "Metadata" : (searchLocationList.SelectedIndex == 2) ? "Files" : "Records";
             btnFind.Text = $"Search {searchLocationList.Text}";
             if(searchLocationList.SelectedIndex == 2) {
-                EntitiesListView.EntityListView.Visible = false;
-                EntitiesListView.SplitContainerToolbar.Panel1Collapsed = true;
-                EntitiesListView.SolutionsDropDown.SolutionType = CrmActions.SolutionType.Unmanaged;
-                EntitiesListView.SolutionsDropDown.LoadData();
+                this.EntityListView.Visible = false;
+                this.SplitContainerToolbar.Panel1Collapsed = true;
+                var items = (List<ListDisplayItem>)this.SolutionDropDownComboBox.DataSource;
+                if (items != null)
+                {
+                    //Filter the solution list
+                    items = items.Where(i => !((Entity)i.Object).GetAttributeValue<bool>("ismanaged") && i.Name.ToLower() != "active" && i.Name.ToLower() != "default").ToList();
+                    this.SolutionDropDownComboBox.DataSource = items;
+
+                }
+                //this.SolutionsDropDown.SolutionType = CrmActions.SolutionType.Unmanaged;
                 this.groupBox1.Text = "Solutions";
             }
             else
             {
-                EntitiesListView.EntityListView.Visible = true;
-                EntitiesListView.SplitContainerToolbar.Panel1Collapsed = false;
-                EntitiesListView.SolutionsDropDown.SolutionType = CrmActions.SolutionType.Both;
-                EntitiesListView.SolutionsDropDown.LoadData();
+                this.EntityListView.Visible = true;
+                this.SplitContainerToolbar.Panel1Collapsed = false;
+                //this.SolutionsDropDown.SolutionType = CrmActions.SolutionType.Both;
+                this.SolutionsDropDown.LoadData();
                 this.groupBox1.Text = "Entities";
             }
         }
