+++ packages.config
@@ -1,7 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="AjaxMin" version="5.14.5506.26202" targetFramework="net472" />
   <package id="DockPanelSuite" version="3.0.6" targetFramework="net472" />
   <package id="DockPanelSuite.ThemeVS2015" version="3.0.6" targetFramework="net472" />
+  <package id="Humanizer.Core" version="2.2.0" targetFramework="net472" />
+  <package id="ICSharpCode.Decompiler" version="7.2.1.6856" targetFramework="net472" />
   <package id="ILMerge" version="3.0.29" targetFramework="net472" />
   <package id="Microsoft.CrmSdk.CoreAssemblies" version="9.0.2.21" targetFramework="net472" />
   <package id="Microsoft.CrmSdk.Deployment" version="9.0.2.21" targetFramework="net472" />
@@ -11,17 +14,28 @@
   <package id="Microsoft.IdentityModel" version="7.0.0" targetFramework="net471" />
   <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="5.2.4" targetFramework="net472" />
   <package id="Microsoft.Web.Xdt" version="3.0.0" targetFramework="net472" />
+  <package id="Microsoft.Win32.Registry" version="5.0.0" targetFramework="net472" />
   <package id="MscrmTools.Xrm.Connection" version="1.2019.11.27" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net472" />
   <package id="NuGet.Core" version="2.14.0" targetFramework="net472" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
+  <package id="System.Collections.Immutable" version="5.0.0" targetFramework="net472" />
   <package id="System.IO" version="4.3.0" targetFramework="net472" />
   <package id="System.IO.Compression.ZipFile" version="4.3.0" targetFramework="net472" />
+  <package id="System.Memory" version="4.5.4" targetFramework="net472" />
   <package id="System.Net.Http" version="4.3.4" targetFramework="net472" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
+  <package id="System.Reflection.Metadata" version="5.0.0" targetFramework="net472" />
   <package id="System.Runtime" version="4.3.1" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.3" targetFramework="net472" />
+  <package id="System.Security.AccessControl" version="5.0.0" targetFramework="net472" />
   <package id="System.Security.Cryptography.Algorithms" version="4.3.1" targetFramework="net472" />
   <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net472" />
   <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net472" />
   <package id="System.Security.Cryptography.X509Certificates" version="4.3.2" targetFramework="net472" />
+  <package id="System.Security.Principal.Windows" version="5.0.0" targetFramework="net472" />
+  <package id="System.ValueTuple" version="4.3.0" targetFramework="net472" />
   <package id="WeifenLuo.WinFormsUI.Docking" version="2.1.0" targetFramework="net472" />
+  <package id="xrmtb.XrmToolBox.Controls" version="2.2.10" targetFramework="net472" />
   <package id="XrmToolBoxPackage" version="1.2019.11.35" targetFramework="net472" />
 </packages>
